Benchmarks: Doc fix and split deriv and ffprop benchmarks. Moved benchmark folder to root.
