+++ benchmarks/bench_actfs_ffprop/results/output_newactfs.txt
@@ -1,0 +1,145 @@
+g++ -std=c++11 -Wall -O3 -Wall -I/home/ithanil/Resilio/PostMaster/Code/NNVMC/FeedForwardNeuralNetwork/src/ -I/usr/local/include -c *.cpp
+g++ -std=c++11 -Wall -O3 -L/usr/local/lib -I/home/ithanil/Resilio/PostMaster/Code/NNVMC/FeedForwardNeuralNetwork/src -L/home/ithanil/Resilio/PostMaster/Code/NNVMC/FeedForwardNeuralNetwork -Wl,-rpath=/home/ithanil/Resilio/PostMaster/Code/NNVMC/FeedForwardNeuralNetwork -o exe *.o -lffnn -lgsl -lgslcblas
+Rebuilt the executable file
+
+
+Ready to run!
+
+--------------------------------------------------------------------------
+
+
+
+FFPropagate benchmark with 500000 FF-Propagations for lgs activation function.
+====================================================================================
+
+NN structure looks like:
+id_    id_    id_    id_    
+id_    lgs    lgs    id_    
+id_    lgs    lgs           
+       lgs                  
+       lgs                  
+       lgs                  
+
+
+Benchmark results:
+noderivs: 0.126312 seconds
+stdderivs: 1.1947 seconds
+allderivs: 11.6947 seconds
+====================================================================================
+
+
+FFPropagate benchmark with 500000 FF-Propagations for gss activation function.
+====================================================================================
+
+NN structure looks like:
+id_    id_    id_    id_    
+id_    gss    gss    id_    
+id_    gss    gss           
+       gss                  
+       gss                  
+       gss                  
+
+
+Benchmark results:
+noderivs: 0.123824 seconds
+stdderivs: 1.17921 seconds
+allderivs: 11.5783 seconds
+====================================================================================
+
+
+FFPropagate benchmark with 500000 FF-Propagations for id_ activation function.
+====================================================================================
+
+NN structure looks like:
+id_    id_    id_    id_    
+id_    id_    id_    id_    
+id_    id_    id_           
+       id_                  
+       id_                  
+       id_                  
+
+
+Benchmark results:
+noderivs: 0.0690516 seconds
+stdderivs: 1.10979 seconds
+allderivs: 11.5853 seconds
+====================================================================================
+
+
+FFPropagate benchmark with 500000 FF-Propagations for tans activation function.
+====================================================================================
+
+NN structure looks like:
+id_    id_    id_    id_    
+id_    tans    tans    id_    
+id_    tans    tans           
+       tans                  
+       tans                  
+       tans                  
+
+
+Benchmark results:
+noderivs: 0.132218 seconds
+stdderivs: 1.21512 seconds
+allderivs: 11.8459 seconds
+====================================================================================
+
+
+FFPropagate benchmark with 500000 FF-Propagations for sin activation function.
+====================================================================================
+
+NN structure looks like:
+id_    id_    id_    id_    
+id_    sin    sin    id_    
+id_    sin    sin           
+       sin                  
+       sin                  
+       sin                  
+
+
+Benchmark results:
+noderivs: 0.145599 seconds
+stdderivs: 1.2735 seconds
+allderivs: 11.8863 seconds
+====================================================================================
+
+
+FFPropagate benchmark with 500000 FF-Propagations for relu activation function.
+====================================================================================
+
+NN structure looks like:
+id_    id_    id_    id_    
+id_    relu    relu    id_    
+id_    relu    relu           
+       relu                  
+       relu                  
+       relu                  
+
+
+Benchmark results:
+noderivs: 0.0843589 seconds
+stdderivs: 1.12378 seconds
+allderivs: 11.5915 seconds
+====================================================================================
+
+
+FFPropagate benchmark with 500000 FF-Propagations for selu activation function.
+====================================================================================
+
+NN structure looks like:
+id_    id_    id_    id_    
+id_    selu    selu    id_    
+id_    selu    selu           
+       selu                  
+       selu                  
+       selu                  
+
+
+Benchmark results:
+noderivs: 0.103123 seconds
+stdderivs: 1.16237 seconds
+allderivs: 11.6603 seconds
+====================================================================================
+
+
+
